Be fix/#128 fix (#138)
* fix/#128: 마일스톤 deadline 형식을 yyyy-MM-dd로 변경

* fix/#128: 마일스톤 생성 시 열림닫힌 상태값 추가하기

* fix/#128: 마일스톤 목록 조회 시 열림/닫힘 상태 구분하기

* fix/#128: 마일스톤 진행상황 정수로 변경하기

* fix/#128: 레이블 목록 반환시 메타데이터 추가

* fix/#128: 마일스톤 목록 반환시 메타데이터 추가

* fix/#128: 메타데이터 추가로 인한 테스트 변경

+++ BE/src/main/java/com/issuetracker/milestone/application/dto/IssueDetailMilestoneInformation.java
@@ -16,7 +16,7 @@ public class IssueDetailMilestoneInformation {
 
 	private Long id;
 	private String title;
-	private Double progress;
+	private Integer progress;
 
 	public static IssueDetailMilestoneInformation from(Milestone milestone) {
 		if (Objects.isNull(milestone)) {
